pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokebenchmark-emulator"
version = "0.1.0"
description = "GBA emulator wrapper and Pokemon Gen 3 game adapters (FireRed, Emerald)"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Shardul Aggarwal" }]
dependencies = [
    "pygba>=0.1.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
include = ["pokebenchmark_emulator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
